Add files via upload
The input of this application is a direcotry of compiled PE files (.exe) and a DLL name. The application parses each PE file to reach out to it's IAT (import address table) and find a list of loaded DLLs.. Next, the application shows a list of DLLs and functions of each DLL that has been used by the loaded PE file, then, if a specific DLL was included in the PE file then the application detects it as a malware.